REsolve problem with the SEQ number in csma802.15.4

+++ src/underTest/linklayer/ieee802154/mac/csma802154.cc
@@ -977,7 +977,8 @@ void csma802154::handleLowerMsg(cMessage *msg)
         else
         {
             //long SeqNr = macPkt->getSequenceId();
-            long SeqNr = macPkt->getBdsn();
+            //long SeqNr = macPkt->getBdsn();
+            uint8_t SeqNr = macPkt->getBdsn();
 
             if (strcmp(macPkt->getName(), "CSMA-Ack") != 0)
             {
